enhance RequestTask
