+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/Task.java
@@ -25,26 +25,6 @@ import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.*;
  */
 public interface Task<C extends Client,F extends Configuration<C,R>,R extends Request> {
 
-
-    /**
-     * Selects a configuration based on the provided request.
-     * If no matching configuration is found, it returns the default configuration.
-     *
-     * @param request the request to match against configurations
-     * @return the selected configuration
-     */
-    default F selectConfiguration(R request) {
-        Preconditions.checkNotNull(request, "Request must not be null.");
-        Map<String,F> configurations = getConfigurations();
-        Preconditions.checkArgument(!configurations.isEmpty(), "Configurations list must not be null or empty.");
-        //is there a matching configuration against the request ?
-        F configuration = getDefaultConfiguration();
-        return configurations
-                .values().stream()
-                .filter(config -> config.matches(request))
-                .findFirst().orElse(configuration); //default configuration
-    }
-
     Map<String,F> getConfigurations();
